[dyad] wrote 4 file(s)
